requirements.txt
ccxt==4.4.85
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
flask==3.0.3